+++ fmriprep/interfaces/images.py
@@ -8,11 +8,11 @@ Image tools interfaces
 
 
 """
-from __future__ import print_function, division, absolute_import, unicode_literals
 
 import os
 import numpy as np
 import nibabel as nb
+import nilearn.image as nli
 
 from niworkflows.nipype import logging
 from niworkflows.nipype.utils.filemanip import fname_presuffix
@@ -22,7 +22,6 @@ from niworkflows.nipype.interfaces.base import (
 from niworkflows.nipype.interfaces import fsl
 from niworkflows.interfaces.base import SimpleInterface
 
-from fmriprep.utils.misc import genfname
 
 LOGGER = logging.getLogger('interface')
 
@@ -62,8 +61,8 @@ class IntraModalMergeInputSpec(BaseInterfaceInputSpec):
 class IntraModalMergeOutputSpec(TraitedSpec):
     out_file = File(exists=True, desc='merged image')
     out_avg = File(exists=True, desc='average image')
-    out_mats = OutputMultiPath(exists=True, desc='output matrices')
-    out_movpar = OutputMultiPath(exists=True, desc='output movement parameters')
+    out_mats = OutputMultiPath(File(exists=True), desc='output matrices')
+    out_movpar = OutputMultiPath(File(exists=True), desc='output movement parameters')
 
 
 class IntraModalMerge(SimpleInterface):
@@ -76,8 +75,8 @@ class IntraModalMerge(SimpleInterface):
             in_files = [self.inputs.in_files]
 
         # Generate output average name early
-        self._results['out_avg'] = genfname(self.inputs.in_files[0],
-                                            suffix='avg')
+        self._results['out_avg'] = fname_presuffix(self.inputs.in_files[0],
+                                                   suffix='_avg', newpath=runtime.cwd)
 
         if self.inputs.to_ras:
             in_files = [reorient(inf) for inf in in_files]
@@ -92,7 +91,8 @@ class IntraModalMerge(SimpleInterface):
                 if sqdata.ndim == 5:
                     raise RuntimeError('Input image (%s) is 5D' % in_files[0])
                 else:
-                    in_files = [genfname(in_files[0], suffix='squeezed')]
+                    in_files = [fname_presuffix(in_files[0], suffix='_squeezed',
+                                                newpath=runtime.cwd)]
                     nb.Nifti1Image(sqdata, filenii.get_affine(),
                                    filenii.get_header()).to_filename(in_files[0])
 
@@ -124,81 +124,214 @@ class IntraModalMerge(SimpleInterface):
         return runtime
 
 
-class ConformSeriesInputSpec(BaseInterfaceInputSpec):
-    t1w_list = InputMultiPath(File(exists=True), mandatory=True,
-                              desc='input T1w images')
+CONFORMATION_TEMPLATE = """\t\t<h3 class="elem-title">Anatomical Conformation</h3>
+\t\t<ul class="elem-desc">
+\t\t\t<li>Input T1w images: {n_t1w}</li>
+\t\t\t<li>Output orientation: RAS</li>
+\t\t\t<li>Output dimensions: {dims}</li>
+\t\t\t<li>Output voxel size: {zooms}</li>
+\t\t\t<li>Discarded images: {n_discards}</li>
+{discard_list}
+\t\t</ul>
+"""
+
+DISCARD_TEMPLATE = """\t\t\t\t<li><abbr title="{path}">{basename}</abbr></li>"""
 
 
-class ConformSeriesOutputSpec(TraitedSpec):
-    t1w_list = OutputMultiPath(exists=True, desc='output T1w images')
+class TemplateDimensionsInputSpec(BaseInterfaceInputSpec):
+    t1w_list = InputMultiPath(File(exists=True), mandatory=True, desc='input T1w images')
+    max_scale = traits.Float(3.0, usedefault=True,
+                             desc='Maximum scaling factor in images to accept')
 
 
-class ConformSeries(SimpleInterface):
-    input_spec = ConformSeriesInputSpec
-    output_spec = ConformSeriesOutputSpec
+class TemplateDimensionsOutputSpec(TraitedSpec):
+    t1w_valid_list = OutputMultiPath(exists=True, desc='valid T1w images')
+    target_zooms = traits.Tuple(traits.Float, traits.Float, traits.Float,
+                                desc='Target zoom information')
+    target_shape = traits.Tuple(traits.Int, traits.Int, traits.Int,
+                                desc='Target shape information')
+    out_report = File(exists=True, desc='conformation report')
+
+
+class TemplateDimensions(SimpleInterface):
+    """
+    Finds template target dimensions for a series of T1w images, filtering low-resolution images,
+    if necessary.
+
+    Along each axis, the minimum voxel size (zoom) and the maximum number of voxels (shape) are
+    found across images.
+
+    The ``max_scale`` parameter sets a bound on the degree of up-sampling performed.
+    By default, an image with a voxel size greater than 3x the smallest voxel size
+    (calculated separately for each dimension) will be discarded.
+
+    To select images that require no scaling (i.e. all have smallest voxel sizes),
+    set ``max_scale=1``.
+    """
+    input_spec = TemplateDimensionsInputSpec
+    output_spec = TemplateDimensionsOutputSpec
+
+    def _generate_segment(self, discards, dims, zooms):
+        items = [DISCARD_TEMPLATE.format(path=path, basename=os.path.basename(path))
+                 for path in discards]
+        discard_list = '\n'.join(["\t\t\t<ul>"] + items + ['\t\t\t</ul>']) if items else ''
+        zoom_fmt = '{:.02g}mm x {:.02g}mm x {:.02g}mm'.format(*zooms)
+        return CONFORMATION_TEMPLATE.format(n_t1w=len(self.inputs.t1w_list),
+                                            dims='x'.join(map(str, dims)),
+                                            zooms=zoom_fmt,
+                                            n_discards=len(discards),
+                                            discard_list=discard_list)
 
     def _run_interface(self, runtime):
-        import nibabel as nb
-        import nilearn.image as nli
-        from nipype.utils.filemanip import fname_presuffix, copyfile
+        # Load images, orient as RAS, collect shape and zoom data
+        in_names = np.array(self.inputs.t1w_list)
+        orig_imgs = np.vectorize(nb.load)(in_names)
+        reoriented = np.vectorize(nb.as_closest_canonical)(orig_imgs)
+        all_zooms = np.array([img.header.get_zooms()[:3] for img in reoriented])
+        all_shapes = np.array([img.shape for img in reoriented])
 
-        in_names = self.inputs.t1w_list
-        orig_imgs = [nb.load(fname) for fname in in_names]
-        reoriented = [nb.as_closest_canonical(img) for img in orig_imgs]
-        target_shape = np.max([img.shape for img in reoriented], axis=0)
-        target_zooms = np.min([img.header.get_zooms()[:3]
-                               for img in reoriented], axis=0)
+        # Identify images that would require excessive up-sampling
+        valid = np.ones(all_zooms.shape[0], dtype=bool)
+        while valid.any():
+            target_zooms = all_zooms[valid].min(axis=0)
+            scales = all_zooms[valid] / target_zooms
+            if np.all(scales < self.inputs.max_scale):
+                break
+            valid[valid] ^= np.any(scales == scales.max(), axis=1)
 
-        resampled_imgs = []
-        for img in reoriented:
-            zooms = np.array(img.header.get_zooms()[:3])
-            shape = np.array(img.shape)
+        # Ignore dropped images
+        valid_fnames = in_names[valid]
+        self._results['t1w_valid_list'] = valid_fnames.tolist()
 
-            xyz_unit = img.header.get_xyzt_units()[0]
-            if xyz_unit == 'unknown':
-                # Common assumption; if we're wrong, unlikely to be the only thing that breaks
-                xyz_unit = 'mm'
-            # Set a 0.05mm threshold to performing rescaling
-            atol = {'meter': 5e-5, 'mm': 0.05, 'micron': 50}[xyz_unit]
+        # Set target shape information
+        target_zooms = all_zooms[valid].min(axis=0)
+        target_shape = all_shapes[valid].max(axis=0)
 
-            # Rescale => change zooms
-            # Resize => update image dimensions
-            rescale = not np.allclose(zooms, target_zooms, atol=atol)
-            resize = not np.all(shape == target_shape)
-            if rescale or resize:
-                target_affine = np.eye(4, dtype=img.affine.dtype)
-                if rescale:
-                    scale_factor = target_zooms / zooms
-                    target_affine[:3, :3] = np.diag(scale_factor).dot(img.affine[:3, :3])
-                else:
-                    target_affine[:3, :3] = img.affine[:3, :3]
+        self._results['target_zooms'] = tuple(target_zooms.tolist())
+        self._results['target_shape'] = tuple(target_shape.tolist())
 
-                if resize:
-                    # The shift is applied after scaling.
-                    # Use a proportional shift to maintain relative position in dataset
-                    size_factor = (target_shape.astype(float) + shape) / (2 * shape)
-                    # Use integer shifts to avoid unnecessary interpolation
-                    offset = (img.affine[:3, 3] * size_factor - img.affine[:3, 3]).astype(int)
-                    target_affine[:3, 3] = img.affine[:3, 3] + offset
-                else:
-                    target_affine[:3, 3] = img.affine[:3, 3]
+        # Create report
+        dropped_images = in_names[~valid]
+        segment = self._generate_segment(dropped_images, target_shape, target_zooms)
+        out_report = os.path.join(runtime.cwd, 'report.html')
+        with open(out_report, 'w') as fobj:
+            fobj.write(segment)
 
-                data = nli.resample_img(img, target_affine, target_shape).get_data()
-                img = img.__class__(data, target_affine, img.header)
+        self._results['out_report'] = out_report
 
-            resampled_imgs.append(img)
+        return runtime
 
-        out_names = [fname_presuffix(fname, suffix='_ras', newpath=runtime.cwd)
-                     for fname in in_names]
 
-        for orig, final, in_name, out_name in zip(orig_imgs, resampled_imgs,
-                                                  in_names, out_names):
-            if final is orig:
-                copyfile(in_name, out_name, copy=True, use_hardlink=True)
+class ConformInputSpec(BaseInterfaceInputSpec):
+    in_file = File(exists=True, mandatory=True, desc='Input T1w image')
+    target_zooms = traits.Tuple(traits.Float, traits.Float, traits.Float,
+                                desc='Target zoom information')
+    target_shape = traits.Tuple(traits.Int, traits.Int, traits.Int,
+                                desc='Target shape information')
+
+
+class ConformOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc='Conformed T1w image')
+
+
+class Conform(SimpleInterface):
+    """Conform a series of T1w images to enable merging.
+
+    Performs two basic functions:
+
+    #. Orient to RAS (left-right, posterior-anterior, inferior-superior)
+    #. Resample to target zooms (voxel sizes) and shape (number of voxels)
+    """
+    input_spec = ConformInputSpec
+    output_spec = ConformOutputSpec
+
+    def _run_interface(self, runtime):
+        # Load image, orient as RAS
+        fname = self.inputs.in_file
+        orig_img = nb.load(fname)
+        reoriented = nb.as_closest_canonical(orig_img)
+
+        # Set target shape information
+        target_zooms = np.array(self.inputs.target_zooms)
+        target_shape = np.array(self.inputs.target_shape)
+        target_span = target_shape * target_zooms
+
+        zooms = np.array(reoriented.header.get_zooms()[:3])
+        shape = np.array(reoriented.shape)
+
+        xyz_unit = reoriented.header.get_xyzt_units()[0]
+        if xyz_unit == 'unknown':
+            # Common assumption; if we're wrong, unlikely to be the only thing that breaks
+            xyz_unit = 'mm'
+
+        # Set a 0.05mm threshold to performing rescaling
+        atol = {'meter': 5e-5, 'mm': 0.05, 'micron': 50}[xyz_unit]
+
+        # Rescale => change zooms
+        # Resize => update image dimensions
+        rescale = not np.allclose(zooms, target_zooms, atol=atol)
+        resize = not np.all(shape == target_shape)
+        if rescale or resize:
+            target_affine = np.eye(4, dtype=reoriented.affine.dtype)
+            if rescale:
+                scale_factor = target_zooms / zooms
+                target_affine[:3, :3] = reoriented.affine[:3, :3].dot(np.diag(scale_factor))
             else:
-                final.to_filename(out_name)
+                target_affine[:3, :3] = reoriented.affine[:3, :3]
 
-        self._results['t1w_list'] = out_names
+            if resize:
+                # The shift is applied after scaling.
+                # Use a proportional shift to maintain relative position in dataset
+                size_factor = target_span / (zooms * shape)
+                # Use integer shifts to avoid unnecessary interpolation
+                offset = (reoriented.affine[:3, 3] * size_factor - reoriented.affine[:3, 3])
+                target_affine[:3, 3] = reoriented.affine[:3, 3] + offset.astype(int)
+            else:
+                target_affine[:3, 3] = reoriented.affine[:3, 3]
+
+            data = nli.resample_img(reoriented, target_affine, target_shape).get_data()
+            reoriented = reoriented.__class__(data, target_affine, reoriented.header)
+
+        # Image may be reoriented, rescaled, and/or resized
+        if reoriented is not orig_img:
+            out_name = fname_presuffix(fname, suffix='_ras', newpath=runtime.cwd)
+            reoriented.to_filename(out_name)
+        else:
+            out_name = fname
+
+        self._results['out_file'] = out_name
+
+        return runtime
+
+
+class ReorientInputSpec(BaseInterfaceInputSpec):
+    in_file = File(exists=True, mandatory=True,
+                   desc='Input T1w image')
+
+
+class ReorientOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc='Reoriented T1w image')
+
+
+class Reorient(SimpleInterface):
+    """Reorient a T1w image to RAS (left-right, posterior-anterior, inferior-superior)"""
+    input_spec = ReorientInputSpec
+    output_spec = ReorientOutputSpec
+
+    def _run_interface(self, runtime):
+        # Load image, orient as RAS
+        fname = self.inputs.in_file
+        orig_img = nb.load(fname)
+        reoriented = nb.as_closest_canonical(orig_img)
+
+        # Image may be reoriented
+        if reoriented is not orig_img:
+            out_name = fname_presuffix(fname, suffix='_ras', newpath=runtime.cwd)
+            reoriented.to_filename(out_name)
+        else:
+            out_name = fname
+
+        self._results['out_file'] = out_name
 
         return runtime
 
@@ -256,8 +389,8 @@ class ValidateImage(SimpleInterface):
 class InvertT1wInputSpec(BaseInterfaceInputSpec):
     in_file = File(exists=True, mandatory=True,
                    desc='Skull-stripped T1w structural image')
-    epi_ref = File(exists=True, mandatory=True,
-                   desc='Skull-stripped EPI reference image')
+    ref_file = File(exists=True, mandatory=True,
+                    desc='Skull-stripped reference image')
 
 
 class InvertT1wOutputSpec(TraitedSpec):
@@ -269,11 +402,9 @@ class InvertT1w(SimpleInterface):
     output_spec = InvertT1wOutputSpec
 
     def _run_interface(self, runtime):
-        from nilearn import image as nli
-
         t1_img = nli.load_img(self.inputs.in_file)
         t1_data = t1_img.get_data()
-        epi_data = nli.load_img(self.inputs.epi_ref).get_data()
+        epi_data = nli.load_img(self.inputs.ref_file).get_data()
 
         # We assume the image is already masked
         mask = t1_data > 0
@@ -291,24 +422,15 @@ class InvertT1w(SimpleInterface):
 
 
 def reorient(in_file, out_file=None):
-    import nibabel as nb
-    from fmriprep.utils.misc import genfname
-    from builtins import (str, bytes)
-
+    """Reorient Nifti files to RAS"""
     if out_file is None:
-        out_file = genfname(in_file, suffix='ras')
-
-    if isinstance(in_file, (str, bytes)):
-        nii = nb.load(in_file)
-    nii = nb.as_closest_canonical(nii)
-    nii.to_filename(out_file)
+        out_file = fname_presuffix(in_file, suffix='_ras', newpath=os.getcwd())
+    nb.as_closest_canonical(nb.load(in_file)).to_filename(out_file)
     return out_file
 
 
 def _flatten_split_merge(in_files):
-    from builtins import bytes, str
-
-    if isinstance(in_files, (bytes, str)):
+    if isinstance(in_files, str):
         in_files = [in_files]
 
     nfiles = len(in_files)
@@ -323,13 +445,13 @@ def _flatten_split_merge(in_files):
             all_nii.append(nii)
 
     if len(all_nii) == 1:
-        LOGGER.warn('File %s cannot be split', all_nii[0])
+        LOGGER.warning('File %s cannot be split', all_nii[0])
         return in_files[0], in_files
 
     if len(all_nii) == nfiles:
         flat_split = in_files
     else:
-        splitname = genfname(in_files[0], suffix='split%04d')
+        splitname = fname_presuffix(in_files[0], suffix='_split%04d', newpath=os.getcwd())
         flat_split = []
         for i, nii in enumerate(all_nii):
             flat_split.append(splitname % i)
@@ -340,24 +462,21 @@ def _flatten_split_merge(in_files):
         merged = in_files[0]
     else:
         # More that one in_files - need merge
-        merged = genfname(in_files[0], suffix='merged')
+        merged = fname_presuffix(in_files[0], suffix='_merged', newpath=os.getcwd())
         nb.concat_images(all_nii).to_filename(merged)
 
     return merged, flat_split
 
 
 def _gen_reference(fixed_image, moving_image, out_file=None):
-    import numpy
-    from nilearn.image import resample_img, load_img
-
     if out_file is None:
-        out_file = genfname(fixed_image, suffix='reference')
-    new_zooms = load_img(moving_image).header.get_zooms()[:3]
+        out_file = fname_presuffix(fixed_image, suffix='_reference', newpath=os.getcwd())
+    new_zooms = nli.load_img(moving_image).header.get_zooms()[:3]
     # Avoid small differences in reported resolution to cause changes to
     # FOV. See https://github.com/poldracklab/fmriprep/issues/512
-    new_zooms_round = numpy.round(new_zooms, 3)
-    resample_img(fixed_image, target_affine=numpy.diag(new_zooms_round),
-                 interpolation='nearest').to_filename(out_file)
+    new_zooms_round = np.round(new_zooms, 3)
+    nli.resample_img(fixed_image, target_affine=np.diag(new_zooms_round),
+                     interpolation='nearest').to_filename(out_file)
     return out_file